+++ crawler/hdu/hdu.py
@@ -396,7 +396,7 @@ class HDUCrawler(BaseCrawler):
 
                     submission_id = cols[0].text.strip()
                     problem_link = urljoin(self.base_url, cols[2].find("a")["href"])
-                    problem_name = cols[2].text.strip()
+                    problem_id = cols[2].text.strip()
                     submission_link = urljoin(self.base_url, cols[5].find("a")["href"])
 
                     # Status format: "Accepted", "Wrong Answer", "Time Limit Exceeded", "Runtime Error (ACCESS_VIOLATION)", etc.
@@ -422,7 +422,7 @@ class HDUCrawler(BaseCrawler):
                         "memory": memory,
                         "language": language,
                         "submit_time": submit_time.isoformat(),
-                        "problem_name": problem_name,
+                        "problem_id": problem_id,
                         "problem_link": problem_link,
                         "submission_link": submission_link,
                     }